requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiosqlite>=0.20.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-mcp-adapters>=0.1.0
lark-oapi>=1.4.0
